+++ kurtosis-module/impl/prelaunch_data_generator/prelaunch_data_generator_launcher.go
@@ -10,7 +10,10 @@ import (
 
 
 const (
-	image = "kurtosistech/ethereum-genesis-generator:0.1.4"
+	// Though this is a Kurtosis image, it's actually built from the original repo:
+	//  https://github.com/skylenet/ethereum-genesis-generator
+	// It's only a Kurtosis image because the original repo doesn't publish Docker images
+	image = "kurtosistech/ethereum-genesis-generator:skylenet-commit-410404d"
 
 	serviceId services.ServiceID = "prelaunch-data-generator"
 )